+++ src/singleTable/implementation.ts
@@ -17,104 +17,14 @@ import {
   SingleTableDeleteParams,
   SingleTableTransactConfigGenerator,
 } from './adaptor';
-import { SingleTableSchema as FullSingleTableSchema } from './model';
-import { SingleTableFromEntity } from './fromEntity';
-import { SingleTableFromCollection } from './fromCollection';
-
-type SingleTableSchema<SingleParams extends SingleTableParams> = {
-  /**
-   * Create a repository-like structure from a Collection to easily extract it joined.
-   *
-   * This could be useful to easily extract item collections and join them.
-   *
-   * Let's say you have a Kanban app in which the following entries are present in your DB:
-   *
-   * - PROJECT
-   * - TAGS
-   * - TASKS
-   * - TASK SUBTASKS
-   * - TASK TAGS
-   *
-   * All of which are under the same PROJECT#id partitionKey
-   *
-   * Create its Collection and the relevant methods by passing it to this method
-   * ```
-   */
-  fromCollection: SingleTableFromCollection<SingleParams>['fromCollection'];
-
-  /**
-   * Create a repository-like structure from an entity, easily execute CRUD methods on said entity
-   *
-   * Every key param will be inferred and queries will be available to directly call
-   */
-  fromEntity: SingleTableFromEntity<SingleParams>['fromEntity'];
-
-  /**
-   * Create a Collection to easily showcase your entities relations and define pre-joined extractions
-   */
-  createCollection: FullSingleTableSchema<SingleParams>['createCollection'];
-
-  /**
-   * Define a Partition more than 1 of your entities share
-   *
-   * Example: You might have an USER partition with a similar use case:
-   *
-   * ```ts
-   *  const partition = table.schema.createPartition({
-   *     name: 'USER_PARTITION',
-   *     getPartitionKey: ({ userId }: { userId: string }) => ['USER', userId],
-   *     entries: {
-   *       data: () => [`#DATA`],
-   *       permissions: () => [`#PERMISSIONS`],
-   *       loginAttempts: ({ timestamp }: { timestamp: string }) => ['LOGIN_ATTEMPT', timestamp],
-   *     },
-   *   });
-   * ```
-   *
-   * By creating a partition, you centralize each Key getter and can easily create the relevant entities out of it
-   */
-  createPartition: FullSingleTableSchema<SingleParams>['createPartition'];
-
-  /**
-   * Define an Entity of your database
-   *
-   * This can be thought of similar to a TABLE on a regular relational database
-   *
-   * To ensure type-safety and its checks, this method requires you to pass the Entity Type
-   * to it, and only then subsequently call `as` to actually execute the creation.
-   *
-   * ```ts
-   *   const User = table.schema.createEntity<UserType>().as({
-   *     // ... entity creation params
-   *   })
-   * ```
-   */
-  createEntity: FullSingleTableSchema<SingleParams>['createEntity'];
-
-  /**
-   * A **low level** wrapper to retrieve the entity registered entity for a given type
-   *
-   * Normally this would be necessary if you want to perform generic operations not knowing ahead of time
-   * what type you have
-   *
-   * Example: Say you have a dynamoDB stream handler that fixes any index partition that may be out-of-sync due to a partial update not containing all
-   * the properties for a composite index key (if key if status+date and you provide the status, the index won't be updated correctly)
-   *
-   * You could then use the stream to read the _type of the modified type, get its entity and check for index presence + values
-   */
-  getEntityByType: FullSingleTableSchema<SingleParams>['getEntityByType'];
-};
+import { SingleTableSchema as FullSingleTableSchema, SingleTableSchemaType } from './model';
 
 export class SingleTable<SingleParams extends SingleTableParams> {
   private methods: SingleTableMethods<SingleParams>;
 
   private fullSchema: FullSingleTableSchema<SingleParams>;
 
-  public schema: SingleTableSchema<SingleParams>;
-
-  private entityRepo: SingleTableFromEntity<SingleParams>;
-
-  private collectionRepo: SingleTableFromCollection<SingleParams>;
+  public schema: SingleTableSchemaType<SingleParams>;
 
   config: Omit<SingleParams, 'dynamodbProvider'>;
 
@@ -125,18 +35,14 @@ export class SingleTable<SingleParams extends SingleTableParams> {
 
     this.fullSchema = new FullSingleTableSchema(params);
 
-    this.entityRepo = new SingleTableFromEntity(params);
-    this.collectionRepo = new SingleTableFromCollection(params);
-
     this.schema = {
       createEntity: this.fullSchema.createEntity.bind(this.fullSchema),
       createPartition: this.fullSchema.createPartition.bind(this.fullSchema),
       createCollection: this.fullSchema.createCollection.bind(this.fullSchema),
       getEntityByType: this.fullSchema.getEntityByType.bind(this.fullSchema),
-
-      fromCollection: this.collectionRepo.fromCollection.bind(this.collectionRepo),
-
-      fromEntity: this.entityRepo.fromEntity.bind(this.entityRepo),
+      from: this.fullSchema.from.bind(this.fullSchema),
+      fromCollection: this.fullSchema.fromCollection.bind(this.fullSchema),
+      fromEntity: this.fullSchema.fromEntity.bind(this.fullSchema),
     };
   }
 
@@ -179,10 +85,17 @@ export class SingleTable<SingleParams extends SingleTableParams> {
     return this.methods.ejectTransactParams(configs);
   }
 
+  /**
+   * [Deprecated soon] Prefer the more clean `transaction`
+   */
   async executeTransaction(
     configs: (SingleTableTransactionConfig<SingleParams> | null)[],
   ): Promise<void> {
-    return this.methods.executeTransaction(configs);
+    return this.methods.transaction(configs);
+  }
+
+  async transaction(configs: (SingleTableTransactionConfig<SingleParams> | null)[]): Promise<void> {
+    return this.methods.transaction(configs);
   }
 
   generateTransactionConfigList<Item extends AnyObject>(