+++ src/singleTable/index.ts
@@ -1,8 +1,5 @@
 export * from './adaptor';
 
-export * from './fromCollection';
-export * from './fromEntity';
-
 export * from './model';
 
 export * from './implementation';