+++ src/singleTable/model/from/fromCollection/utils/entityMap.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { ExtendableCollection, ExtendableSingleTableEntity } from 'singleTable/model';
+import type { ExtendableCollection, ExtendableSingleTableEntity } from 'singleTable/model';
 
 export type EntityMap = Record<string, ExtendableSingleTableEntity>;